requirements.txt
seaborn
pandas
matplotlib.pyplot
matplotlib.image 